+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Leaflet Changelog
 
 ## Unreleased Changes
 These are on the `main` or `develop` branch and will be included in the next numbered release.
-
+- BREAKING: Change `L.motion.seq.motionStart()` to behave the same as `L.motion.polygon.motionStart()` (do nothing if currently animating)
 ## 0.3 (2021-09-17)
 - Added CHANGELOG.md to track future changes ([@patabugen](https://github.com/patabugen))
 - Added CONTRIBUTING.md to help other people get involved ([@patabugen](https://github.com/patabugen))